% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMCMCmaxlik.R
\name{buildMCMCmaxlik}
\alias{buildMCMCmaxlik}
\title{Construct compiled functions for finding MLE.}
\usage{
buildMCMCmaxlik(model, paramNodes)
}
\arguments{
\item{model}{a nimble model object}

\item{paramNodes}{a character vector indicating the top-level parameters}
}
\description{
This function builds a list of compiled functions for finding the MLE.
}
\examples{
buildMCMCmaxlik()
}
\keyword{MLE}

