% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMLE.R
\name{computeMLE}
\alias{computeMLE}
\title{Compute the MLE given a BUGS model}
\usage{
computeMLE(model, paramNodes, method = "fixed", paramInit = NULL,
  compiledFuns = NULL, ...)
}
\arguments{
\item{model}{a nimble model object}

\item{paramNodes}{a character vector indicating the top-level parameters}

\item{method}{a string indicating the optimization strategy}

\item{paramInit}{the initial starting point}

\item{compiledFuns}{a list of compiled function, generated by buildMCMCmaxlik()}
}
\description{
This function computes the MLE of the top-level parameters of a
hierarchical model with latent variables.
}
\examples{
compileMLE()
}
\keyword{MLE}

